// Frontend/ClinicaIntegra/src/app/models/MedicoModel.ts
export interface MedicoModel {
  COMUNA_id_comuna: number;
  ROL_id_rol: number;
  apellido_materno: string;
  apellido_paterno: string;
  correo: string;
  direccion: string;
  fecha_nacimiento: string;
  nombre: string;
  nombre_cargo: string;
  nombre_especialidad: string;
  personaRun: string;
  telefono: string;
}